backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
openai-whisper==20240930
transformers==4.46.3
torch==2.5.1
librosa==0.10.2